refact: adds depreacations
