+++ lib/capybara_screenshot_diff/rspec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "rspec/core"
-require "capybara/screenshot/diff"
+require "capybara_screenshot_diff"
 
 module Capybara::Screenshot::Diff
   ASSERTION = ::StandardError unless defined?(::Capybara::Screenshot::Diff::ASSERTION)
